Fix for empty ConfirmSectionLabware.
If a user forgets to fill in any section numbers on the sectioning confirm screen,
the SectioningConfirmMachine never builds its list of ConfirmSections that are sent
to stan-core (as the build is triggered by a user action).

Fix is to build the ConfirmSections when the machine starts instead of waiting for
a user event.